Add bashrc_local reading and protect CERN Root
bashrc now sources ~/.bashrc_local, if it exists. bashrc_local is not
included in this repository, because it should be the settings that are
unique to each machine.

The function root_check() is now protected by an if statement to see if
the executable exists. If it does not, root is never aliased, so local
versions of root should still work.