Kill pidgin when calling pidgin-chmod
We kill pidgin when calling pidgin-chmod so that log files are closed
before being marked read only. If the files were not closed and then
synced to another system with pidgin-push/pull, the push/pull script
would fail the next time it was run because it would be unable to modify
the files.