Change vnoremap to xnoremap in vimrc
vnoremap remaps in visual and select mode, xnoremap remaps only in
visual mode. Select mode is slightly different in that if you type a
single character vim replaces the selected area with the character (this
mimics the behavior of, for example, Microsoft Word when you select an
area of text). Remapping characters in select mode makes this behavior
inconsistent.